requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
whitenoise==6.6.0
PyPDF2==3.0.1
python-docx==1.1.2
openai==1.82.0
